(feat&fix) completed update task feature
Fix the bug that didn't let to pass the information from the father process to the instance. Now update both instances and task when a task is completed.
Fixed the route /getTaskByUserGroup to be a get for a more correct http verb usage. Now uses a String that is converted to ObjectId to search for the correct user group